+++ Runtime/DialogueDatas/Events/DialogueEvent_ImageColor.cs
@@ -1,0 +1,41 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace DSC.DialogueSystem
+{
+    [CreateAssetMenu(fileName = "DialogueEvent_ImageColor", menuName = "DSC/Dialogue/Events/Image Color")]
+    public class DialogueEvent_ImageColor : DialogueEvent
+    {
+        #region Variable
+
+        #region Variable - Inspector
+#pragma warning disable 0649
+
+        [SerializeField] int m_nIndex;
+        [SerializeField] Color m_hColor;
+
+#pragma warning restore 0649
+        #endregion
+
+        #endregion
+
+        #region Base - Override
+
+        public override void OnStart(List<IDialogueEventData> lstData)
+        {
+            base.OnStart(lstData);
+
+            if (!lstData.TryGetData(out DialogueEventData_Image hOutData))
+                return;
+
+            var hImageGroupController = hOutData.m_hGroupController;
+            if (hImageGroupController == null)
+                return;
+
+            hImageGroupController.SetImageColor(m_nIndex, m_hColor);
+        }
+
+        #endregion
+    }
+}